LP1274999: Next Link Appears at "End" of Copy List
The total number of copies wasn't checked against the
copy_limit and copy_offset, so anytime that the
number of copies per page mod copy_limit == 0 the
Next link would appear, even if there were only
copy_limit copies total. Now the offset and limit are
compared to the total number of copies available at
the current depth to determine if the Next link
should appear.

Signed-off-by: Jason Boyer <jboyer@library.in.gov>
Signed-off-by: Jason Etheridge <jason@esilibrary.com>