LP#1669522: Remove regex that caused duplication in author query
In LP#1308090, a regex replacement was added to the normzalize_qterm block
that didn't work as intended. Instead of replacing the already-normalized qterm,
it added a new string to the qterm, causing the author and date terms to
duplicate in the resulting query when the link is clicked. Since the the
problem the regex was trying to solve appears to be an aesthetic one rather
than a functional one, this commit removes that regex.

Signed-off-by: Kathy Lussier <klussier@masslnc.org>
Signed-off-by: Chris Sharp <csharp@georgialibraries.org>
Signed-off-by: Ben Shum <ben@evergreener.net>