LP#1522644: Make Mark for Hold Transfer option consistent with other options
Makes the following changes to the Mark for Holds Transfer Destination option:

1) Changes the label from "Mark As Title Hold Transfer Destination" to "Title
Hold Transfer." The 'Mark" piece is already implied in the "Mark For" button
label, and the long text was pushing the menu off the screen for some displays.

2) Makes the option behave as other Mark options. If a record is already marked,
it displays the marked Record ID in the menu. The mark is also removed when the
"Reset Record Marks" option is clicked.

3) Pushes the reset option down to the bottom of the menu and alphabetizes the
other Mark options above it.

Signed-off-by: Kathy Lussier <klussier@masslnc.org>
Signed-off-by: Bill Erickson <berickxx@gmail.com>