LP#1673857: teach catalog how to search and display copy tags
When the opac.search.enable_bookplate_search library setting is
set to true, the catalog will display a "Digital Bookplates" search
field in the drop-downs on both the search bar and the advanced search
page. Using this will add a "copy_tag(*, search_terms)" filter
to the search, i.e., all visible copy tags will be searched regardless
of type.  Users can also use the copy_tag() search filter directly.

Visible copy tags are displayed in the copy table in the record
summary page. Note that copy tags are displayed regardless of whether
or not opac.search.enable_bookplate_search is on or off.

Mike Rylander also contributed to this patch.

Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>
Signed-off-by: Josh Stompro <stomproj@larl.org>
Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>