LP#1680624 Consolidate package dependencies into package.json
npm already enables us to designate dependencies in package.json,
and projects have increasingly consolidated what bower used to be
used for in package.json. We can do this with an npm-compatible version of
angular-order-object-by from github.

There are a few other tweaks to npm vs bower file locations, such as
ng-toast instead of ngtoast for ngToast.

Pin to AngularJS 1.5.x for now

Signed-off-by: Dan Scott <dan@coffeecode.net>
Signed-off-by: Bill Erickson <berickxx@gmail.com>
Signed-off-by: Ben Shum <ben@evergreener.net>