LP#1257915 Repair receive-when-cancel query
Repaire the json_query used to determine if a PO is ready to be marked
received.

A PO is non-receiveable if it has any lineitems that are not in the
received/cancelled [sic] state OR any that are canceled with a
keep_debits=true cancel reason.

Prior to this change, simply having a state of "cancelled" was enough to
prevent receiving.

* Replace tabs w/ spaces
* Remove unnecessary "or return 0" clause which was causing Perl
  precedence warnings.

Signed-off-by: Bill Erickson <berickxx@gmail.com>
Signed-off-by: Chris Sharp <csharp@georgialibraries.org>
Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>