LP#1680554: fix upgrade of subject|temporal_browse and subject|geographic_browse indexes
This patch fixes an issue where an Evergreen database that was
upgraded to 2.12.0 from an earlier version may not have had
the browsability of the subject|temporal and subject|geographic
indexes turned off in favor of the new *_browse versions. It also
fixes mapping of the new *_browse indexes to authority control
set bib fields.

Note that if the database started at 2.12.0, the browse
reingest can be cancelled.

To test
-------
[1] On a database that was upgraded to 2.12.0 from an earlier
    version, verify that after applying the database update and
    performing the browse reingest, that there are no
    metabib.browse_entry_def_map rows where def is 11 or 13. Also
    Also verify that the query that runs as part of the update
[2] On a database that was installed starting at 2.12.0, verify
    that the database update is a no-op.

Signed-off-by: Kathy Lussier <klussier@masslnc.org>