LP#1673857: add ability to set copy tags in volume/copy editor
The copy editor now has a 'Copy Tags' button that can be used
to assign or remove tags from a copy. A typeahead widget is
used to allow the user to select an existing tag, but users can
also use this interface to create an entirely new tag on the fly.

Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>
Signed-off-by: Josh Stompro <stomproj@larl.org>
Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>