LP1690434: Date Comparisons in the Reporter
Because of the way the where clause is constructed
in queries using the Year + Month Transform reports
using some types of operators could produce invalid
SQL.

To test:
put together a simple report template with a date
in the Basic Filters section, using the Year + Month
transform and the On or After operator. (a count of
circs with an xact_start filter is a simple way to
set this up.) Schedule it with a relative date
such as 1 month ago and run it.

Pre Patch: the report will fail and the error text
will reference a missing text >= double precision
comparison function.

Post Patch: you'll get a count of how many circs
were opened last month (or whatever report you
wrote).

Signed-off-by: Jason Boyer <jboyer@library.in.gov>
Signed-off-by: Mike Rylander <mrylander@gmail.com>