LP#1677902: Remove search formats from advanced search limiter block
The original behavior of the advanced search limiter block was to remove
search formats from the display because it is available for limiting in the
search bar. Other code broke this behavior, and I've now restored it by
adding the ignore search format code to the searchbar.tt2 file. It also fixes
a problem where the search format limiters were incorrectly displaying on the
low hits page.

Signed-off-by: Kathy Lussier <klussier@masslnc.org>
Signed-off-by: Jeff Davis <jdavis@sitka.bclibraries.ca>