LP#1642042 Webstaff noncat real-time display
Update non-cat counts in the patron summary side bar and show the full
set of non-cat circs in the Non-Cataloged Circulations tab as non-cat
circs occur in the patron checkout interface.

Consistent with the XUL client, avoid updating tab-level 'items out'
counts with non-cat circs.

Adds a new work log action for 'noncat_checkout' so the correct data can
be extracted.  As with the XUL client, this appears in the work log as a
'checkout' with no copy.

Signed-off-by: Bill Erickson <berickxx@gmail.com>
Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>