Docs: Fix buggy AsciiDoc rendering of backticks
Apparently, AsciiDoc matches backticks (`) with single quotes (') as a
valid quote pair, and this happens before matching a pair of backticks.
So in cases where a pair of backticks is followed by only one single
quote (not within a word), the first backtick is treated as a single
quote, thus breaking the backtick pair and not applying the intended
monospace formatting.

This commit fixes two such cases by escaping the backticks. This doesn't
prevent the same problem in the future, but I haven't found any way to
change this parsing behavior.

Signed-off-by: Remington Steed <rjs7@calvin.edu>