LP#1677000 - webstaff Adds a 'has-notes visual indicator' for patrons with notes.
Made the indicator be clickable and route to Other->Notes.
Other ways to address this not mentioned in LP would be
using a glyphicon instead, or moving the Notes submenu item out of Others tab.
Or even making the entire tab/menu be resposive, and collapsing into an Others submenu
only when needed.

Signed-off by: Cesar Velez <cesar.velez@equinoxinitiative.org>

Signed-off-by: Jason Etheridge <jason@equinoxinitiative.org>