LP#1693560: Remove permission check for retrieve tag table function
The permission check for the retrieve tag table function was preventing
users who don't have the UPDATE_MARC permssion from performing searches
in the staff catalog. Since there's nothing confidential in the table,
this commit removes that permission check. Thanks to Jason Boyer for
finding the source of the problem.

Signed-off-by: Kathy Lussier <klussier@masslnc.org>
Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>