LP#1669907: Web Staff Bib record Holds View dups
In the webstaff client, switching from the View Holds tab
to the OPAC View and back would result in the list of hold
requests getting duplicated.

This patch fixes the problem.

Caching of tab data is explicitly not being done ATM,
so this takes care of the issue for now.

Signed-off-by: Cesar Velez <cesar.velez@equinoxinitiative.org>
Signed-off-by: Terran McCanna <tmccanna@georgialibraries.org>
Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>