LP#1673870: Use separate templates for ebook checkouts and holds
Previously we used the same TT2 template for listing checkouts and
performing a checkout, and for listing holds and placing a hold.  This
leads to problems; in particular, URL parameters required for checkout
and hold actions were "sticky," so viewing your holds after placing a
hold would show the "Place Hold" UI instead of a list of holds.  This
commit introduces separate templates for ebook actions as a workaround.

Signed-off-by: Jeff Davis <jdavis@sitka.bclibraries.ca>
Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>
Signed-off-by: Bill Erickson <berickxx@gmail.com>