LP1702911: Correct Barcode State in Add Item
When adding an item from the MARC editor the barcode
state was incorrect because the empty_barcode flag
is set on the copy and then not changed after
assigning a barcode. At that point the only way to
alter the flag is modifying the barcode in the UI,
which the user likely would not think to do. This
causes the Save buttons to remain disabled.

To test pre patch:
Open a record in the MARC editor, check Add Item
Add a call number label and barcode above, click Save
In the vol/copy editor, try as you might, you can't save. :(

To test post-patch:
Open a record in the MARC editor, check Add Item
Add a call number label and barcode above, click Save
In the vol/copy editor, the Save buttons will already be enabled
(barring required stat cats, etc.)

Signed-off-by: Jason Boyer <jboyer@library.in.gov>
Signed-off-by: Josh Stompro <stomproj@larl.org>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>