LP#1486451: Remove rdetails_status nowrap style
The copy details table was styled to not wrap, which sometimes pushed table
contents off the page, particularly when some of the copies are conjoined
items. This commit removes the style that was preventing wrapping.

Test plan:
Load record 15 in the Concerto dataset. Before the patch is loaded, you'll
see that some of the content in the copy details table is pushed off the
screen and requires horizontal scrolling. Post-patch, all of the copy
details will display on screen without horizontal scrolling. In addition
to the change in the Conjoined Items, you also might see some text wrapping
in the Shelving Location and Status columns.

Signed-off-by: Kathy Lussier <klussier@masslnc.org>
Signed-off-by: Terran McCanna <tmccanna@georgialibraries.org>
Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>