LP 1189989: Add suspend option when placing hold
This commit adds some bells and whistles to the basic functionality:

1. Adds two new styles in style.css.tt2 for toggled blocks and inline
text.  These styles allow text with id attributes beginnging with
certain strings to be normally hidden and then shown when they are the
target of an active anchor.  Only the block form is currently used,
but the inline text version may prove useful in the near future.

2. Moves the "Suspend this hold?" language above the checkbox and adds
a variant to make it clear that all of the holds will be suspended
when multiple holds are placed.

3. Adds the question mark graphic with mouseover help text indicating
that a suspended hold will not be fulfilled until it has been
activated.

4. Adds a "Set activation date" anchor to toggle the appearance of the
text input for the optional thaw date.

5. Adds a script block with a JavaScript function to toggle the
appearance of the text input for the optional thaw date and to check
the suspend hold check box when first toggled.  This function returns
false to prohibit the href of the anchor from firing and toggling the
appearance via CSS.  If JavaScript is disabled, the input's appearance
will be toggled via CSS, but the checkbox will not be checked
automatically.

6. Adds language to the place hold results page to indicate if the
hold is suspended and adds the reactive date if supplied.

7. Modifies the backend holds code to not set the expire time on a
hold if it is suspended.  This is a logical extension of code added
for Lp bug #1076399: Activating an expired hold cancels the hold.

PLEASE NOTE: I would very much like to keep this branch in three
commits because the first commit is Bill Ott's initial implementation
from 2013.  The second commit is a fix to Bill's work required to make
it work with modern ttopac.  This third commit overs the expansion of
the functionality to meet MassLNC requirements.  I think it would
behoove us to keep these three pieces of work in distinct commits.

Signed-off-by: Jason Stephenson <jason@sigio.com>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>
Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>