LP#1635737 Use new OpenSRF interval_to_seconds() context
Use the optional context for interval_to_seconds() to account for the
variable length of duration components.  For example, "1 day" may be
shorter or longer than 24 hours during a time change event, "1 month"
may be shorter or longer depending on which month it is currently, etc.

Also, remove some timestamp munging, as that happens within
interval_to_seconds() already.

Signed-off-by: Dan Wells <dbw2@calvin.edu>
Signed-off-by: Mike Rylander <mrylander@gmail.com>

Conflicts:
	Open-ILS/src/perlmods/lib/OpenILS/Application/Circ/Circulate.pm

Signed-off-by: Dan Wells <dbw2@calvin.edu>
Signed-off-by: Bill Erickson <berickxx@gmail.com>