LP#1706365: cursor for record menu items should be a pointer
When using the Mark For menu on the bib record page, the
cursor becomes an i-beam while hovering over the menu items instead of the
expected pointer. This commit adds an empty href attribute to the a tag for menu
items so that a pointer displays instead.

Test plan:
Log into the web client and retrieve a record. Click the 'Mark For' menu on the
record page. Pre-patch, as you scroll through the menu items, the cursor will
turn into an i-beam, indicating that text can be selected. Post-patch, the
cursor will be the expected pointer that typically displays for links.

Signed-off-by: Kathy Lussier <klussier@masslnc.org>
Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>