LP#1635737 Apply DST-aware timezone to context dates
Do our best to enforce the rule required by OpenSRF's interval_to_seconds
that when a context date is in use, and you care about DST awareness, you
must set the timezone to a DST-aware value, e.g., 'America/New_York'. In
most situations, 'local' will suffice for this, as the server is typically
configured with a DST-aware timezone in its environment.  However, we will
look for an org unit setting called 'lib.timezone' and use that where we
can.  See LP#1705524 for info on that setting.

Signed-off-by: Mike Rylander <mrylander@gmail.com>

Conflicts:
	Open-ILS/src/perlmods/lib/OpenILS/Application/Circ/Circulate.pm
	Open-ILS/src/perlmods/lib/OpenILS/Application/Circ/NonCat.pm

Signed-off-by: Dan Wells <dbw2@calvin.edu>
Signed-off-by: Bill Erickson <berickxx@gmail.com>