LP#1189989: (follow-up) ignore invalid thaw date
This patch ensures that the specified thaw date is valid,
as otherwise specifying a thaw date of "13/08/2017" (when,
at present, only dates of the form mm/dd/yyyy are accepted)
would result in an Apache internal server error.

If an invalid date is supplied, the thaw date is ignored, meaning
that the hold request will be placed as suspended but will not
automatically reactivate.

Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>