LP 1694058: Add backend code to allow multiple hold placement.
We add a constant for the circ.holds.max_duplicate_holds setting.

We modify Holds.pm to check if we're placing a title or metarecord
hold, that we have the CREATE_DUPLICATE_HOLDS permission, and that we
haven't placed more than the maximum allowed number of duplicate holds
before returning the HOLD_EXISTS event.

Signed-off-by: Jason Stephenson <jason@sigio.com>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>